pyproject.toml
[tool.poetry]
name = "2022-ABPS-fNIRS-Demo"
version = "0.1.0"
description = "Demonstrate fNIRS processing with NIRSport2 device."
authors = ["Robert Luke <code@robertluke.net>"]

[tool.poetry.dependencies]
python = "3.8.13"
click = "^8.1.3"
pyxid2 = "^1.0.4"
PyYAML = "^6.0"
mne = "^1.0.3"
mne-nirs = "^0.2.1"
mne-qt-browser = "^0.3.1"
ipython = "^8.4.0"
jupyterlab = "^3.4.3"
ipykernel = "^6.15.0"
seaborn = "^0.11.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
